finanalyst/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1
pdfplumber==0.11.0
openpyxl==3.1.2
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
python-dotenv==1.0.1
reportlab==4.1.0
joblib==1.3.2
pydantic==2.6.4
